Rewritten how instructionCounter works.

+++ src/ALU/ALUMemoryHandler.cpp
@@ -1,5 +1,5 @@
 #include "../../include/ArithmeticLogicUnit.h"
-int ArithmeticLogicUnit::sc_memorySet(int address, short value)
+int ArithmeticLogicUnit::sc_memorySet(int address, int value)
 {
     if ((address >= 0) && (address <= 99)) {
         if (value <= 0x3FFF && value >= -0x4000) {
@@ -12,7 +12,7 @@ int ArithmeticLogicUnit::sc_memorySet(int address, short value)
     sc_regSet(OUT_OF_MEMORY, 1);
     return -1;
 }
-int ArithmeticLogicUnit::sc_memoryGet(int address, short& value)
+int ArithmeticLogicUnit::sc_memoryGet(int address, int& value)
 {
     if ((address >= 0) && (address <= 99)) {
         value = memory[address];
